requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
neo4j==5.17.0
Jinja2==3.1.3 
openai==1.82.0
pydantic-settings==2.9.1